refactor: Rename Navbar to temp_navbar
