+++ src/pages/profile/Profile.tsx
@@ -1,18 +1,16 @@
 import React, { useEffect } from "react";
 import { MainContainer } from "@shared/ui/mainContainer";
-import { Navbar } from "@widgets/navbar";
+import { Navbar } from "@/widgets/temp_navbar";
 import { AddShelter } from "@widgets/addShelter/index.ts";
-import * as petModel from '@entities/pet/index.ts';
+import * as petModel from "@entities/pet/index.ts";
 export const Profile: React.FC = () => {
   const { data: favorites } = petModel.api.useGetFavoritesQuery();
-  useEffect(()=>{
-  
-  },[favorites])
+  useEffect(() => {}, [favorites]);
   return (
     <>
       <MainContainer>
         <h1>PROFILE</h1>
-        <AddShelter/>
+        <AddShelter />
       </MainContainer>
       <Navbar />
     </>
