+++ src/widgets/temp_navbar/ui/Navbar.tsx
